refactor(ui): decoupe config-page.tsx en douze composants
Le fichier passe de 1962 a 54 lignes et ne fait plus que composer. Les
douze composants vivent dans src/ui/config/, organises par section de
formulaire puis par panneau lateral.

Le rendu est identique octet pour octet : meme SHA-256 sur GET /, 76807
octets avant et apres, idem pour /logs, /api/openapi.json et /llms.txt.

Cette preuve n'etait pas une formalite. Une premiere passe dedentait les
blocs extraits, ce qui cassait l'indentation a l'interieur des <pre> du
panneau doc, ou elle est du contenu et non de la mise en forme. Le diff a
revele 8 octets d'ecart dans un exemple curl. Le texte brut est desormais
decoupe sans jamais etre reindente, deno fmt ne touchant pas a l'interieur
des gabarits.

Invariants verifies sur le HTML final : les cinq fieldset ont leur legend
en enfant direct, aucune reference aria orpheline sur 96 identifiants, et
les 77 identifiants exiges par assertElement existent tous.

Neuf constantes Tailwind couvrant 40 occurrences, limitees aux classes
structurelles des controles de formulaire. La prose du panneau doc reste en
place : en extraire les classes obligerait a aller les lire pour savoir a
quoi ressemble un dt, ce qui degraderait la lisibilite.
